fixed error subscription fixed error average rating

+++ src/main/java/app/wooportal/server/components/push/FirebasePushService.java
@@ -42,29 +42,34 @@ public class FirebasePushService {
   }
 
   public void sendPush(UserEntity user, MessageDto message) {
-    for (var iterator = user.getSubscriptions().iterator(); iterator.hasNext();) {
-      var subscription = iterator.next();
-      try {
-        var messageBuilder =
-            Message.builder().setToken(subscription.getDeviceToken()).setNotification(Notification
-                .builder().setTitle(message.getTitle()).setBody(message.getContent()).build());
+    if (user.getSubscriptions() != null) {
+      for (var iterator = user.getSubscriptions().iterator(); iterator.hasNext();) {
+        var subscription = iterator.next();
+        try {
+          var messageBuilder =
+              Message.builder().setToken(subscription.getDeviceToken()).setNotification(Notification
+                  .builder().setTitle(message.getTitle()).setBody(message.getContent()).build());
 
-        if (message.getData() != null) {
-          messageBuilder.setAndroidConfig(AndroidConfig.builder().putAllData(message.getData()).build())
-              .setApnsConfig(ApnsConfig.builder()
-                  .setAps(Aps.builder().putAllCustomData(new HashMap<String, Object>(message.getData()))
-                      .build())
-                  .build())
-              .setWebpushConfig(WebpushConfig.builder()
-                  .setNotification(WebpushNotification.builder().setData(message.getData()).build())
-                  .build());
+          if (message.getData() != null) {
+            messageBuilder
+                .setAndroidConfig(AndroidConfig.builder().putAllData(message.getData()).build())
+                .setApnsConfig(ApnsConfig.builder()
+                    .setAps(Aps.builder()
+                        .putAllCustomData(new HashMap<String, Object>(message.getData())).build())
+                    .build())
+                .setWebpushConfig(
+                    WebpushConfig.builder()
+                        .setNotification(
+                            WebpushNotification.builder().setData(message.getData()).build())
+                        .build());
+          }
+          FirebaseMessaging.getInstance().sendAsync(messageBuilder.build()).get();
+
+        } catch (InterruptedException | ExecutionException e) {
+          e.printStackTrace(System.out);
+          iterator.remove();
+          subscriptionService.deleteById(subscription.getId());
         }
-        FirebaseMessaging.getInstance().sendAsync(messageBuilder.build()).get();
-
-      } catch (InterruptedException | ExecutionException e) {
-        e.printStackTrace(System.out);
-        iterator.remove();
-        subscriptionService.deleteById(subscription.getId());
       }
     }
